added data to stock allocation
