+++ src/components/input/BranchAutoComplete.tsx
@@ -24,8 +24,18 @@ export default function BranchAutoComplete(props: {
   );
 
   useEffect(() => {
-    // Ensure that the selected value is set when the component is mounted
     if (props.value) {
+      let selectedValue: number | null = null;
+
+      if (typeof props.value === "string") {
+        const numericValue = parseFloat(props.value);
+
+        if (!isNaN(numericValue)) {
+          selectedValue = numericValue;
+        }
+      } else {
+        selectedValue = props.value;
+      }
       const selectedBranch = filteredBranch?.find(
         (branch: any) => branch?.BPLID === props.value
       );
@@ -35,7 +45,6 @@ export default function BranchAutoComplete(props: {
     }
   }, [props.value, filteredBranch]);
 
-  // Use local state to store the selected value
   const [selectedValue, setSelectedValue] = useState(null);
 
   const handleAutocompleteChange = (event: any, newValue: any) => {
