+++ src/presentations/retail_sale/fuel_cash_sale/components/NozzleDataTable.tsx
@@ -46,7 +46,6 @@ export default function NozzleData({ data, onChange, edit }: NozzleDataProps) {
       ];
 
   console.log(datattest);
-
   if (tl_Dispenser.length > 0) {
     data.nozzleData = tl_Dispenser;
   }
